AT45DB flash-interface: set b_addr of at45db_read_buffer() to uint16_t

+++ platform/inga/interfaces/flash-at45db.c
@@ -194,7 +194,7 @@ void at45db_page_to_buf(uint16_t addr) {
 
 }
 
-void at45db_read_buffer(uint8_t b_addr, uint8_t *buffer, uint16_t bytes) {
+void at45db_read_buffer(uint16_t b_addr, uint8_t *buffer, uint16_t bytes) {
 	uint16_t i;
 	uint8_t cmd[4] = { AT45DB_READ_BUFFER, 0x00, (uint8_t) (b_addr >> 8),
 			(uint8_t) (b_addr) };
